requirements.txt
# Core Dependencies
python-dotenv==1.2.1
requests==2.32.5

# Google Gemini API
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2

# Streamlit UI
streamlit>=1.28.0

# Data Validation
pydantic==2.12.5
pydantic_core==2.41.5

# Utilities
tqdm==4.67.1
cachetools==6.2.2

# HTTP & Security
certifi==2025.11.12
charset-normalizer==3.4.4
httplib2==0.31.0
idna==3.11
urllib3==2.6.2
rsa==4.9.1

# Protocol Buffers
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2

# Type Checking
annotated-types==0.7.0
typing-inspection==0.4.2
typing_extensions==4.15.0

# Parsing
pyparsing==3.2.5
uritemplate==4.2.0
